refactor: Update tag module structure and imports
