--- project/libs/blog/blog-tag/src/blog-tag-module/blog-tag.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('blog-tag-controller')
-export class BlogTagController {}
