+++ project/libs/blog/blog-tag/src/index.ts
@@ -3,4 +3,4 @@ export { BlogTagFactory } from './blog-tag-module/blog-tag.factory';
 export { BlogTagModule } from './blog-tag-module/blog-tag.module';
 export { BlogTagRepository } from './blog-tag-module/blog-tag.repository';
 export { BlogTagService } from './blog-tag-module/blog-tag.service';
-export { TagRdo } from './rdo/tag.rdo';
+export { TagRdo } from './blog-tag-module/rdo/tag.rdo';
